requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.5
django-redis==5.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.2
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-http-client==3.3.7
redis==5.2.1
requests==2.32.3
sendgrid==6.11.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==11.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
